requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
gunicorn
pyjwt
pwdlib[argon2]